Use custom hook
